requirements.txt
torch
fastapi
uvicorn
pinecone-client
python-dotenv
